+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Fixed a rare bug, where sometimes MOP would not have access to the save file
 - Fixed not attached fuel tank getting toggled on and off repeatedly
 - Fixed "mop rules" command displaying the info about the rules incorrectly
+- Fixed a bug, in which MOP would fail to locate correct Satsuma key object
 
 ## 3.9.2 (07.10.2022)